pyproject.toml
[project]
name = "overleaf-cookie"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "nopecha",
    "playwright>=1.50.0",
    "python-dotenv>=1.0.1",
]

[tool.uv.sources]
nopecha = { git = "https://github.com/r0-negative/nopecha-python.git" }
